Towards parallelisation: BackwardStaticCondenstation + DistributedDarcyHDGTests
